[NNAPI] Remove solid weights support
Summary:
This was left over from an old version of the code.  The idea was that
instead of indexing into separate tensors for each weight, you could
bundle them all into a single file and use different offsets into that
file.  With the current design, this is nontrivial to support, so drop
the code for now.

Test Plan:
CI

ghstack-source-id: ddcc5673f5e8b6b5ce06e3a8c59ac70e89c7eb4f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47518