remove hacky_wrapper from BackendSelect
BackendSelect kernels have been changed to be written the new way, so this hacky_wrapper here isn't needed anymore.
This PR is not expected to change perf or anything, just simplify the code a bit. The hacky_wrapper here was a no-op and not creating any actual wrappers
because it short-cirtuits to not create a wrapper when there is no wrapper needed.

Differential Revision: [D25421633](https://our.internmc.facebook.com/intern/diff/D25421633/)

ghstack-source-id: 118177101
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49079