hacky_wrapper_for_legacy_signatures reorders out arguments
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48911

This enables us to use hacky_wrapper_for_legacy_signatures for ops with out arguments so they can use templated unboxing logic without having to be rewritten.

This only actually enables it for one op as a proof of concept. There will be a separate PR enabling it for more ops.
ghstack-source-id: 118181603

Differential Revision: [D25363336](https://our.internmc.facebook.com/intern/diff/D25363336/)